feat(eval_sim): prop-eval optimal-control solvers + reclasso empty-active-set fix
Frame the prop-eval as goal-reaching-before-drawdown (an up-and-in/down-and-out
barrier option) and solve for the optimal position-sizing control:
- hjb_pde.solve_pde: implicit backward-Euler + Rannacher + Howard policy iteration
  on the running-max reduction, positive-coefficient (viscosity-safe). hjb.py is the
  explicit reference DP (superseded for accuracy; kept for the derivation).
- mc_fast: antithetic + process-parallel first-passage MC.
- adp: geometric block-bootstrap env, risk-constrained-Kelly (Busseti-Ryu-Boyd) anchor,
  and the LITERATURE-optimal StructuredPolicy (Yuan & Li 2021; Browne 1999): Merton
  fraction far from the barrier, bang-bang at the leverage cap (bold) near it -- press
  when behind. Beats the anchor +0.043 (paired DM p~1e-253). fit_adp (LASSO fitted-Q
  via the reclasso enet) kept as the data-driven refinement (loses to the closed form
  here: the optimum is a sharp bang-bang a smooth Q cannot represent).
- reclasso_har: fix empty-active-set float-dtype IndexError in _homotopy_walk
  (np.asarray(a, dtype=int); no-op when non-empty, so bit-exactness is preserved).

+++ reclasso_har.py
@@ -56,8 +56,12 @@ def _homotopy_walk(
     s: list[float] = [float(np.sign(c[a[0]]))]
     just_dropped = -1
     for _ in range(max_iter):
-        a_arr = np.asarray(a)
-        s_arr = np.asarray(s, dtype=float)
+        a_arr = np.asarray(
+            a, dtype=int
+        )  # dtype=int so an EMPTY active set stays integer-indexable
+        s_arr = np.asarray(
+            s, dtype=float
+        )  # (else np.asarray([]) -> float64 -> np.ix_ IndexError)
         g_aa = g[np.ix_(a_arr, a_arr)]
         w0 = _solve(g_aa, c[a_arr])
         w1 = _solve(g_aa, s_arr)
